requirements.txt
python-binance>=1.0.19
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0